requirements.txt
sounddevice>=0.4.6
soundfile>=0.12.0
numpy>=1.24.0
websockets>=12.0
groq>=0.4.0
gtts>=2.5.0
pydub>=0.25.1
python-dotenv>=1.0.0
openai>=1.0.0
